pyproject.toml
[tool.poetry]
name = "vadim_docs"
version = "0.1.0"
description = ""
authors = ["Alexey Baranov <atlantij@gmail.com>"]

[tool.poetry.scripts]
run-vadim = 'vadim_docs.cli:cli'

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.47.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
gsheets = "^0.6"
pyOpenSSL = "^22.0.0"
simple-settings = "^1.2.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


